Improve examples, improve variance section and address some comments
This commit rewrites most of the examples to be more realistic. It also
rewrites the variance section to be more explanatory. A note is added
describing variance annotations. A small section mentioning polymorphic
variant subtyping is also added. It includes other small changes to
address some of the comments from the beta.

closes ocamllabs/rwo-comments#1941
closes ocamllabs/rwo-comments#1940
closes ocamllabs/rwo-comments#1878
closes ocamllabs/rwo-comments#1875
closes ocamllabs/rwo-comments#1859
closes ocamllabs/rwo-comments#1805
closes ocamllabs/rwo-comments#1769
closes ocamllabs/rwo-comments#1768
closes ocamllabs/rwo-comments#1767
closes ocamllabs/rwo-comments#1765
closes ocamllabs/rwo-comments#1759
closes ocamllabs/rwo-comments#1758
closes ocamllabs/rwo-comments#1620
closes ocamllabs/rwo-comments#1576