Rewrite the classes chapter.
In particular, reorder some of the sections and change all the examples. This
includes removing the cryptokit example section (as it did not really use
classes). I have also added an "open recursion" section to explain the
"object (self)" syntax and to emphasise open recursion as the key feature of
classes.

There are now three overarching examples: a stack (continued as an example from
the objects chapter), a class to fold over a document format, and some
interactive graphical shapes (using async_graphics) that are constructed using
mixins.