let pygments control background
We had set background to transparent for all <pre> nodes, overriding
the setting in _pygments.scss. Remove this on main code blocks, so
pygments styles with background will show the background.

However, code blocks within notes and asides are still made transparent.

Finally, had to move the padding settings from the <pre> to the
enclosing <div>. Otherwise, the padding part wouldn’t have the
background applied to it.