opam: remove camlp4 constraint
camlp4 4.02.1+2 seems to no longer be in opam, which is probably good
because other camlp4 versions don’t order correctly by opam. For
example, 4.02+6 is considered earlier when it is actually a later
release. In any case, opam now picks up a version of camlp4 that is
okay for issue #112, which is the reason we had added this constraint.