add a `rwo eval` command and make code dir moveable
the eval command will generate a sexp of the script part which is
cached, and use that if available. This lets the code repo never
be evaluated for html generation.

Also make it easier to move the code examples to a separate directory
since it now lives in github.com/realworldocaml/examples