Reduce horizontal coverage for clickable area in links
Previously, the clickable portion of these links was the entire horizontal width of their container instead of just the part of the link that has visible text.  While browsing the book, I mis-clicked a few times on what I thought was empty space, but was actually a ToC or Navbar link.

`display: table-cell` is a great replacement for `display: block` because it keeps all the vertical layout the same, while also not expanding to fill its available space horizontally.