Add dependency over textwrap and yojson in concurrent-programming
This is not ideal as we'd like the libraries from the duniverse
to be picked up without that trick but we'd need to rethink
how we organize the repo as we can't achieve that if the examples
or marked as ignored/data-only. We could mark them as vendored
although that's not strictly what they are. To be discussed further.