remove unused discourse js and fix syntax highlighting
the PrismJS we are using has a small bug in the CLI plugin
(which we use https://prismjs.com/plugins/command-line/)
that doesnt escape/unescape HTML from the output

This PR upgrades PrismJS to 1.17 and also adds a one-line fix
that does escape the output.  In particular, the guided tour
is now consistent with the output of utop, since the `<fun>`
output from the toplevel doesn't corrupt the HTML any more.

Closes #3079