Work around dune 2's read only targets in PDF generation
We just moved to dune 2 to be able to use the mdx stanza and as a side
effect it broke the PDF generation because targets are read-only in dune
2.x.
Some of the PDF generation rules seem to overwrite the `book.idx`, which
doesn't work in dune 2.x.

As a temporary work-around I sequenced the problematic rules into a
`progn` so that the intermediate files can be tempered with as much
as we want.

It would be good to investigate this in the future!