Example rules: pick dune in PATH
Previously, example rules would invoke dune using `(run dune ...)`.

When seeing `(run prog ...)`, dune first looks for a public executable
named `prog` in the workspace before looking at the environment
(`PATH`).

In the case of this repository, it happens that a dune executable is in
the workspace, because the sources of dune-build-info are there and they
are the same as the dune sources.

When building the examples, this dune (2.6 at the moment) will be
picked, and it will try to read dune-package files for the other
packages. If these have been written by a dune in the environment with a
higher version, it will fail with an error. Effectively, this puts a
constraint that dune in the environment should be older than the
vendored one.

To fix this, the `(run dune ...)` action is replaced by `(system "dune
...")`. This is treated as an opaque string passed to the system shell,
which resolves using the environment only. This causes the vendored dune
to be ignored.