GADT Chapter: Permit the naive variant implementation to compare bools
The desired module type signature has this signature for `eq`

```ocaml
val if_ : bool t -> 'a t -> 'a t -> 'a t
```

But the naive implementation throws an exception if it receives anything other than two ints.  This isn't really necessary, so I propose a change to the program which allows it.